deploy/docker_images/configurator/requirements.txt
mysql-connector==2.1.4
PyYAML==5.3.1
SQLAlchemy==1.1.4
pymongo==3.5.1
pika==1.1.0
influxdb==4.1.1

# our dependencies
config-client==1.0.77
